[XTENSA] HIGHMEM  - DEBUG Kernel ran last night appeared pretty stable.
This is my 2nd HIGHMEM check-in; It's now appears to be relatively stable,
and seems like a good point for getting more feedback and to start merging
over to the A1000 tree; a new A2000 board with 1GB was delivered last
Friday. I'm puting this on a new branch, trying to make it possible to
delete the old preliminary commit in the future as it had a number of
thing wrong and likely wouldn't help in understanding HIGHMEM. git may
require a merge and force it's being kept in the git history.

The current version DOESN'T have any support for cache aliasing.
The ARM code seems to have addressed it recently and I thought I'd
update the highmem code once it's solid on a simple DC233L; which
doesn't have aliases. I've started moving to more of the ARM HIGHMEM
code, for example making use of a new generic HIGHMEM interface
I've assimilated from 2.6.37.

I've removed much of DEBUG code that I'll be removing
more as it seem prudent.

I updated the Kconfig to have HIGHMEM selects SMP, without
SMP Hardware (MX). This was for a few reasons thought it now
appears that at least the page_alloc code is likely not effected
as I first thought. Making the use of MX hardware to be independent
of SMP is more flexible and possibly improved stability.


I've hacked some of the generic kernel code temporally to ease debugging
and possibly fixing a few problems:

	1. Page Group Mobility was disabled, at least prior to updating
	   update_mmu_cache() it had a big improvement in avoiding list
           corruption in page_alloc.c

	2. Assimilated some ARCH_NEEDS_KMAP_HIGH_GET code from the
           the generic mm/highmem.c to facilitate getting the page
	   mapping like ARM is currently doing; seems better than
	   my previous code.

Kernels compiled with CONFIG_DEBUG_KERNEL, without optimization,
with lots of memory usage ran for 13 hours last evening before
I stopped it. Stability seems significantly improved.

I've updated the comments in entry.S TLB miss handler as I made
sure I understood the code 100.00%. It's not squeezed into 80 columns;
but I value clarity more than dogma/convention; at least for now.

Added boot parameter parsing to a few files; it's more conventional;
for example the bootmem code. Used this convention for new highmem code.

I'll check in the current A2000 code into a git repo and then merge
this code with it to provide an environment for initial testing.

I updated the .xt-gdbinit file with breakpoints useful for debugging
this HIGHMEM code.

I check in a copy of my debug .config file I'm currently using into
the arch/xtensa/configs directory.

Tried to document everything so others can easily work with the
new environment.

Current code also runs when HIGHMEM isn't configured. Haven't tested
with HIGHMEM and EXTENDED_MEMORY together; but don't expect any problems.

Signed-off-by: Pete Delaney <piet@tensilica.com>